fix: creating checks and business rules for the transaction

+++ api/models/transaction.go
@@ -6,7 +6,7 @@ func ValidateTransaction(transaction Transaction, user User, wallet Wallet) erro
 
 type Transaction struct {
 	Model
-	Value float64 `gorm:"default:0.00" json:"value"`
+	Value float32 `gorm:"default:0.00" json:"value"`
 	Payer uint64  `gorm:"not null" json:"payer"`
 	Payee uint64  `gorm:"not null" json:"payee"`
 }
